Modularize the build system
Considering the inclusion of more graphical backends and image loaders,
it is vital to make the build system modular so that contributors can
focus on the newly added parts rather than having to deal with build
rules and dependencies. This commit introduces some helpers for basic
management of source/target files based on GNU Make, simplifying the
existing rules. However, the configurations are currently hardcoded.