Tweak build messages for dependency generation
The 'cpp' (C Preprocessor) tool is used for dependency tracking for
C/C++ files. However, showing the message 'CPP' might confuse someone
into thinking that the build system is executing the C++ compiler
instead of the C Preprocessor, as per the original meaning of 'cpp'.
Currently, the message is tweaked to 'DEP', implying the generation of
dependency tracking.