Make build system fully configurable
With the inclusion of kconfiglib [1], the build system is now fully
configurable like "menuconfig" when building the Linux kernel. It allows
users to select the desired backend, choose whether to build
applications, and specify supported image loaders. This flexibility is
also useful for developers to integrate additional backends such as
Linux framebuffer and DRM device support.

[1] https://pypi.org/project/kconfiglib/