Manipulate mouse cursor with a minimal image
This commit makes mouse cursor manipulation configurable, allowing it to
be turned off completely for the SDL backend and to be featured in the
upcoming Linux framebuffer backend with a minimal mouse image rendered.
Currently, only one minimalist dot is provided, but the default cursor
image should be improved later.