Describe fixed-point types
In the original comment, the fixed-point type was only
explained using 12.4 and 24.8, lacking details about
the number of bits allocated for the integer, fractional,
and sign parts.

Here, following the supplementary content for C99 in
'section 6.3.6.3 fixed-point types' of the ISO/IEC TR 18037:2008,
additional descriptions of these fixed-point types are provided.

In particular:
In a fixed-point type where
N bits are used to represent the value, with
L bits allocated for the integer part, the remaining
N−L bits are used for the fractional part.
The actual value range of this fixed-point type is from -2^L to 2^{L}-2^{N-L}.

Additionally, a table is provided showing hexadecimal, decimal, binary,
and actual values to help developers quickly understand
the value range of this fixed-point type.

Refs:
	1. ISO/IEC TR 18037:2008
	   4.2 Detailed changes to ISO/IEC 9899:1999