Describe geometric types and function in font-edit
Currently, there is a lack of description for geometric types and functions.

pt_t is used to describe a geometric object—a point, while pts_t is used to
describe a dynamically configurable array.
n is used to describe the current number of configured points,
and s is used to describe the maximum configurable memory limit, measured in bytes.

spline_t is used to describe cubic spline interpolation based on four points.

new_pts, dispose_pts, and add_pt are methods for the pt_t and pts_t types.

distance_to_point, distance_to_line, and distance_to_segment
are used to describe the distance relationships between geometric objects.

lerp is a function for computing linear interpolation points.

fit is used to compute a spline with a specific error tolerance.