kconfig: Add an option to allow empty macros
When parsing Kconfig which include macros it is currently necessary to
provide a value for all macros in advance. This may not be possible in
some cases, e.g. when the caller is performing checks on the Kconfig
options but is not running a full build of the project.

Add an option to support this. This allows parsing of Zephyr Kconfig
files without specifying a particular board, etc.

Signed-off-by: Simon Glass <sjg@chromium.org>