Consolidate logging with a unified interface
Currently, various messages are displayed while the window system is
running, but their format and usage are inconsistent. Additionally,
depending on the scenario, it may be useful to adjust the color scheme
or disable logging output altogether.

The rxi/log.c [1] offers a robust solution for these needs, and it would
be beneficial to integrate it into the core implementation of the window
system to standardize logging and enhance flexibility.
 - log_trace
 - log_debug
 - log_info
 - log_warn
 - log_error
 - log_fatal

[1] https://github.com/rxi/log.c

Close #47