Reduce cursor latency by rendering to VNC framebuf
Previously, cursor movement was handled by passing pointer events into
twin screen, and then calculated the damage region for the cursor's new
position and transmitted it to the VNC framebuffer. This process
introduced latency in cursor display due to the multiple steps involved.
To address this delay, the cursor is now rendered directly within the
VNC framebuffer by leveraging neatvnc's internal cursor API. This change
bypasses the intermediate steps and provides a smoother and more
responsive cursor movement experience in the VNC backend.