Fix memory region in framebuffer
The size of the screen may be different to the virtual resolution. If the size
is less than virtual resolution, it will draw multiple windows on display. Use
`line_length` to calculate instead of the size of the screen.
Changable information like virtual resoultion or setting different information
on virtual terminal may not be applied on different hardware and driver. For
example on VirtualBox, virtual resolution's width and height are 2048 and 2048,
visible resolutions' are 800 and 600, and line length is 2048 * 4. Any setting
from guest machine to these maybe ignored.