Implement the drop shadow
Add the `twin_stack_blur()` function to implement Mario's Stack Blur
algorithm, which blurs the target pixel map. Additionally, create a test
window to evaluate the effect of `twin_stack_blur()`. Implement a shadow
pixel map beneath the window with a blur effect to create a frosted
glass appearance. The drop shadow effect of the window is only visible
when it is on the top layer, ensuring the active window stands out
visually.

Ref: https://melatonin.dev/blog/implementing-marios-stack-blur-15-times-in-cpp/
See #34

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>