Fix display issue and heap-buffer-overflow
The issue with the display not showing the rightmost pixel of the icon
and clock is caused by commit ceb748f. Therefore, in ceb748f, solving
the heap-buffer-overflow issue by using width - 1 is incorrect. We
revert the changes from commit ceb748f and instead perform a check in
`_span_fill()`. The last pixel may have already been covered when
processing the middle pixels. To address this, an additional condition
is added when processing the last pixel to ensure it is not covered by
the middle pixels, allowing us to handle the last pixel separately.

See sysprog21#49
See sysprog21#51
Close sysprog21#78