Fix widgets not updating with parent's widget
Implement the twin_widget_children_paint function to repaint all the
box's widgets. The issue arises when the spline updates the pixmap
beneath the button widget, the pixmap of the button isn't triggered
to repaint. To solve this, we call the twin_widget_paint function after
the spline update to ensure the related widget will be repainted.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>