Fix global redisplay not matching window updates
The global redisplay thread would block on a semaphore, which is
signaled by the per-window redisplay threads whenever the content of any
window changes. For example, the per-window redisplay thread is triggered
when its content twin_box_t is triggered by the event
TwinEventButtonDown. In this case, the corresponding window should also
be marked as active. To ensure proper display, we added the function
twin_window_show to indirectly update the damaged screen.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>