Fix active window not working on the framebuffer
Remove the inefficiently function twin_active_pixmap(). This function
finds the currently and previously active window pixel map by traversing
the list of pixel maps. This function will be called every time the
screen needs to be updated. Set the active variable of the window only
when the window's event TwinEventButtonDown is triggered to avoid
redundant operations.

Close #92

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>