Fix the wrong range of pixel map for pixman
When the rendering is based on Pixman, the title bar and the gaps on
the far right and bottom of the window will be rendered. However, these
regions should not be accessible. Implement a check for the window's
clipping variable to ensure that only the accessible pixel range is
rendered.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>