Change the logic for window click detection
Implement the function twin_window_valid_range() to check whether the
click coordinates fall within the window's range, based on the window's
style. Originally, the logic for detecting if the window was clicked
relied on the function twin_pixmap_transparent(). However, the click
event won't be triggered when a transparent pixel in the window's region
is clicked.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>