Add iconify and restore features on window frame
Add a condition in twin_window_dispatch() to handle clicks on the
iconify and restore buttons located on the window frame. The
positions of these buttons are calculated for click detection.

A iconified window is inactive, even if it remains at the top of the
screen. A window is considered active when it is the topmost window and
is not iconified.

When the window is active, its title bar turns blue.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>